add copies of viewer_*_t.lcm messages from drake
These messages implement the drawing protocol between drake and
drake-visualizer.  The messages are currently in drake, but the plan
is to remove them from there so that they are only in this repository.

Note, these messages are also in bot_core_lcmtypes.  We are waiting on
openhumanoids and drake to begin using this robotlocomotion/lcmtypes repo.